+++ src/IMU/docker-compose.yml
@@ -1,0 +1,37 @@
+version: '2'
+services:
+    odsupercomponent:
+        image: seresearch/2018-dit-168:v0.3.5-armhf
+        network_mode: "host"
+        restart: always
+        volumes:
+        - .:/opt/miniature.data
+        command: "/opt/od4/bin/odsupercomponent --cid=111 --verbose=1 --configuration=/opt/miniature.data/configuration"
+
+    proxy-miniature-pwm-motor:
+        image: seresearch/2018-dit-168:v0.3.5-armhf
+        depends_on:
+        - odsupercomponent
+        network_mode: "host"
+        restart: always
+        privileged: true
+        volumes:
+        - /sys/:/sys/
+        - /dev/:/dev/
+        command: "/opt/opendlv.miniature/bin/opendlv-proxy-miniature-pwm-motor --cid=111 --freq=50 --id=1"
+
+    device-adc-bbblue-left:
+        image: chalmersrevere/opendlv-device-adc-bbblue-armhf:v0.0.6
+        network_mode: "host"
+        privileged: true
+        volumes:
+          - /dev/mem:/dev/mem
+        command: "/usr/bin/opendlv-device-adc-bbblue --cid=111 --freq=5 --channel=2 --id=0"
+    
+    device-adc-bbblue-right:
+        image: chalmersrevere/opendlv-device-adc-bbblue-armhf:v0.0.6
+        network_mode: "host"
+        privileged: true
+        volumes:
+          - /dev/mem:/dev/mem
+        command: "/usr/bin/opendlv-device-adc-bbblue --cid=111 --freq=5 --channel=3 --id=1"
