+++ src/IMU/messages.odvd
@@ -3,3 +3,9 @@ message readingsIMU[id = 2202] {
 	uint8 readingSteeringAngle [id = 2];
 	uint8 readingSpeed [id = 3];	
 }
+
+message opendlv.proxy.AccelerationReading [id = 1030] {
+  float accelerationX [id = 1];
+  float accelerationY [id = 2];
+  float accelerationZ [id = 3];
+}
